new: rptmux: emit keys from ratpoison instead of calling tmux
rptmux originally queried tmux, got a list of its mappings, and then
rebound the mapped key (less the prefix) to super modifier within
ratpoison.  the mechanism used was to take the actual mapped tmux
command, and call that using the tmux command line interface.  however,
this broke when multiple tmux client existed, attached to different
sessions (which we use sometimes to get different terminals with
different font sizes for different things we're working on).  Tmux
command line did not know which session was the target one, so it simply
chose the first one, which only worked if that happened to be what the
current client was connected to.

We could have changed the tmux command to figure out every time which
terminal it was being run from, which could have worked out a client.
Instead we've changed the logic altogether so the ratpoison bindings do
not call tmux, but use its 'meta' command to emit a keypress that
activates tmux to do the right thing.  This way we care not about what
commands are used by tmux, only which keys.  We simply make an alternate
parallel map with Super prefix that emits tmux prefix, followed by the
suffix key.

As before, it still only works with 'prefix' keymap, and with
single-character keybindings in tmux.  This however covers most use
cases for myself, and there are no other known users...