chore: make the test for withdrawn windows happen every time
originally we only did this on a mismatch, but we can increase the
robustness of this code by doing it every time.  originally the purpose
of this was because we found a scenario where the withdrawn windows
were at the end, but we cannot handle that without a whole lot of
code changes which we do not really need to do.  so now that the
code was moved anyways to try to handle that case, leave it there,
since it will catch withdrawn windows better and make more clear
why we care about withdrawn windows