fix: remove index mismatch test that false-matches for edge case
when the withdrawn windows are at the end, it will always mismatch, because
there was never a chance to rewrite $titles[] to remove the faulty withdrawn
member (since it was the $rpdata[] length, i.e. $nwins that governed the
loop iteration count, and we had already stopped by then).  originally we
wanted to do this by running a separate pass over $titles[] to remove all
withdrawn windows, but after thinking about this, it became clear that this
does not really matter, since $rpdata[] is really what we care about, and
it does not ever have withdrawn windows in it.  so we only really need to
ignore withdrawn windows if they are found before we finished looking at
all the windows known to ratpoison.  after we finished that, we no longer
care about any remaining windows, so ignore the mismatch by removing
this test entirely