- xtty: swallow stderr closer to potential errsite
if 'ps' could not find anything it would give a very verbose error,
which we didn't want to expose the user to, so we ran the whole script
under exec 2>/dev/null.  unfortunately this means set -x output will be
swallowed and we cannot debug.  also it calls out the fact that we
probably should have done this closer to the actual site where we wanted
to swallow error.  so do exactly that: only swallow it from the subshell
that runs 'ps', and then test for its emptiness later to tell us it
failed