+ bright: give up/down their own symlink for acpid
acpid can set actions, scripts to run based on acpi events like funkeys

the scripts called by acpid do not allow arguments however.  so we use
symlinks to ferry in an argument by $0.  for this reason we give new
names 'brightup' and 'brightdown'

which means that system acpi daemon will be configured to call us when
installed in ~/bin/ which should be fixed, but right now we have no srcx
for /usr/local/, that shall come at a later time