- x: qtile: don't specify the config file, use the defaults
we can't actually install the default config file using installrc, it's
the first one encountered

the main reason why: it can't start with a dot.  that's a special syntax
in python import and all that the program does is chop off the last 3
chars and feed to python __import__(), which won't work if starts with
dot.

also, there's other reasons... we have -maxdepth 1 on the find, but we
need the link to go into ~/.config/qtile/config.py so no way to give it
all the information it needs without creating directories

this can be fixed later, for now, for the single machine where this wm
is running, will just manually ln -s ../../.qtilerc.py
~/.config/qtile/config.py and that should keep it happy.  but need to
think about this because lots of stuff is switching to using the
~/.config convention (unfortunately...)