+ wmpersist: implement restores from different groups
now that group information is saved in the persist file, we can restore
into groups also.  each window we check about the saved group vs the
current group and move it if it's not in the right group.  this is quite
a hack and we're basically doing a manual script of ratpoison window
movement commands, but it works (the whole script is a hack; we'll do
better once we have switched to qtile)