- wmpersist: increase sleep to avoid ratpoison race when renumbering
seems there is some race and ratpoison ends up losing track of the
window, it exists, and you can switch to it with wmctrl, but ratpoison
doesn't seem to know anything about it, doesn't show in menus, and
moving the window with rpleft/rpright wrappers does nothing.  increasing
this sleep seems to make it work well enough.  no idea what's causing
this race, and we'll switch to qtile soon, and ratpoison is not
developed anymore, so there's no point looking into it further