+ rparrange: replace rpleft/rpright with a python implementation
the python implementation will allow to implement rprenumber more easily
(which was stubbed, but never implemented in rparrange.sh), is
[arguably] easier to follow, and can be extended with other stuff more
easily if needed

we also fixed the rotate off the end, ie rpleft on the first window or
rpright on the last window.  previously these just swapped the first and
last windows.  now it will rotate it truly to the end for left at top,
and will shift the entire list right to make room at top if the last one
is rotated to the right and there's no room