+ rparrange: rpafter: new invoke, adds new window after current one
'rpafter' exec's a given command, and rearranges windows to make the new
one get inserted just after the one that was current when window spawn
was run.  this way we don't have new windows popping up in random
locations

we also have to add another invocation 'rptrigger' because we need to
wait on the new window spawn, and there's no way to know when that
finished, because it happens in the background (ie in a fork() and isn't
waited for by ratpoison).  but we need the new window to get mapped and
ratpoison switch to it before proceeding.  so we work around this by
adding a trigger command and then mapping it to the new window hook, and
waiting for it after we spawn the window.  this uses a fixed named pipe
as the intermediary

to avoid deadlock, we use select(2) to be sure input is waiting, and
make sure we read the correct number of bytes