- rparrange: rpafter: add debug prints, hard to debug in debugger
it's difficult to debug this in the debugger, because changing windows
will destroy the right state transition sequence that the script
expects.  realistically, we can work around this by either logging in
from another computer to do the debug, or by adding prints and watching
the trace from a run after it happened.  we add the relevant debug
prints here, including a utility function that only prints when $DEBUG
is active.

as a side effect we'll also print these debug statements even for other
invocations when they run rp(), but this is fine and useful anyways.  it
was better to do the print in rp() itself rather than before each
invocation of it