/ wmpersist: fix flock(1) usage bug, no persists for last two weeks
this was idiot command line usage.  '-n' is *not* fdnum.  the fdnum is
taken just as an arg on its own.  in fact '-n' means something else
entirely.  this means we didn't save anything for the last two weeks, so
when we accidentally started a paused browser and brave crashed, wham:
we have to peice things back together because we have no recent
persisted browser state