+ rparrange: rpleft: make it able to take a count arg
instead of just 1, let it go as many as requested in the call (defaults
to 1 if not provided)

this doesn't attempt to handle the scroll-over if we give an arg that
results in going past the end.  except in the case of count=1.

TODO we should implement this differently in the future to allow this by
doing a shift on windows[] and then rpnumber with two args to make the
switches without needing to keep track of current window