- bright: teach it about a 'max' arg that sets brightness to max
different machines have different values for brightness_max and
therefore different scales, so the same absolute number is very
different on different machines (some like thinkpad are 1-16, some like
xps are 96,000, probably depends on how fine-grained the hardware
control is available on the underlying device).  we allow specifying
absolute number, but this can get us in a jam if we "forgot" the scale,
we might not be able to type because display is too dim or off.

now, just using 'max' will restore the brightness to the advertised (in
sysfs) max brightness value