+ urxvtc: add DISPLAY to sockpath for multiple same-host instances
otherwise they will use the same socket but actually it's connected to a
different X server...

by default the hostname component is already present, now we are adding
display, but we have to inject hostname and it seems $HOSTNAME is not
available so we have to expand that too

note that the program gives misleading values about the socket name in
lsof, the files that get opened are actually different than the ones
shown in lsof.  it's probably reading a memory buffer, which starts as a
default but changes after config file is read, or something like that.

witness:

  $ s lsof | grep urxvt/urxvtd-omnius | awk '{print $(NF-1)}'
  /home/scott/.urxvt/urxvtd-omnius.smemsh.net

  $ grep urxvtd /proc/net/unix | awk '{print $NF}'
  /home/scott/.urxvt/urxvtd-omnius.smemsh.net:0.0

bottom one is one in config file and is correct one actually being used,
determined by testing.  not sure if this is bug in lsof or in linux or
in urxvt or in something else, but did test two different ones
concurrently and they correctly use different sockets