- gg: only rewrite greys to be bright for rxvt, other terms ok
rxvt makes darker greys really dark.  it's possible that it's just doing
a better map into the 256 color grayscale range than xterm did, but
nonetheless, without mapping to xterm colors (which are brighter)
better, applications like elinks have default grey colors that are too
dark.  this problem doesn't exist on xterm or suckless for example.  so
we'll only do it for TERM ~ rxvt