/ wmpersist: along with rp, make work from cron with sdorfehs
there will not be any SDORFEHS_PID set in cron, so we'll make a second
attempt to see if it's sdorfehs by contacting the X server and getting
WM_NAME from root window (which a suckless tool called wmname already
does).  this is set correctly at least on ratpoison, but even if that
doesn't work then we will fallback to ratpoison

not that we'll ever use ratpoison proper again.  we are on sdorfehs and
did a few patches on top of that, so no going back.  but we will go to
qtile before doing much else anyways