- rptmux: need to handle space although only 1-chars before this
we had a comment even that we only handle one char, but we need to
handle this.  we first just check, if it's anything but a single char,
is it 'space', then ok.  we'll have to maintain this in two places then,
maybe we should just drop the single char check, and try to write out
all of the mappings.  for nowe we'll maintain this dual whitelist and
see how many times we have to update it