+ rptaskline: write all key-invoked taskw/timew events to a logfile
without having the terminal for all task stuff, and also having those in
command line history, we will be losing some out-of-band representation
of the event information.  we still have the timew database, but it
helps to be able to independently correlate it with other things (like
shell command history timestamps).  by writing it and using a date and
one event per line we can at least consider joining it later and it's
nice to have this in one place as a log, besides the timew db and use of
timew queries