/ mousepark: position 5 pixels in from edge, at 0 vtabs won't expand
all the way to the edge (mouse X at 0) apparently is too far to trigger,
we have to move it a few pixels into the window.  it's looking
internally for a mouseover event, it looks like from brave source code.
so, if X coordinate is 0, that's probably considered outside the page.

it's possible 1 might work, but the minimized bar is way more than 5
pixels anyways so this value should be fine