+ rptaskmenu: new invocation rptaskallmenu does not filter out completed
ordinarily this is more useful if we only show open tasks.  we have
added a new flag to 'timerecent', namely '-o/--open' that only will show
tasks that aren't already completed.  here we make that the default, but
leave any other invocation (in this case we have a symlink
'rptaskallmenu' that will do it) will show all tasks, not just open
ones.  intention is to have both as keybinds in ratpoison because both
are useful at certain times.