+ mousepark: unpark: warp to minimum X coordinate for vtab to disappear
if we park it while the mouse is over to the left but not yet to the
edge, then unparking will not move it far enough to the right that the
bar will hide.  it has to go beyond the edge of the bar in order for
auto-hide to activate.  unfortunately this is purely a user-set width by
dragging the bar so we have to hardcode it.

there might be some way to query brave about the current setting for
vbar width, we can look into this later, we set maximum width and are
unlikely to change it with any frequency on a particular workstation.