= mousepark: todo: add [malfunctioning] code that should warp pointer
this would save us from having to use pynput, we only do the one thing.
probably it is calling XWarpPointer itself, we need to look at the args
it's getting to be able to set absolute coordinates.  tried many
variations and it's not working right with any off them so we will keep
pynput for the moment